remove support for loading IMG/CCD/SUB
Suggested by malvarenga123: https://github.com/xjsxjs197/WiiSXRX_2022/issues/139#issuecomment-2075589747

To me that answer was really strange as the sub file is very small compared to the main image, so you shouldn't need to read it that frequently.
What I think is more plausible is that the general reading routine (for all formats) is not really optimized and the constant access to USB or SD card is causing those stutters.